# Improve Kubeturbo discovery performance by caching pod ownerReferences during cluster monitoring