--- a/pkg/action/executor/action_executor.go
+++ b/pkg/action/executor/action_executor.go
@@ -3,11 +3,10 @@ package executor
 import (
 	"github.com/openshift/cluster-api/pkg/client/clientset_generated/clientset"
 	"github.com/turbonomic/kubeturbo/pkg/action/util"
+	"github.com/turbonomic/kubeturbo/pkg/cluster"
 	"github.com/turbonomic/kubeturbo/pkg/resourcemapping"
 	"github.com/turbonomic/turbo-go-sdk/pkg/proto"
 	api "k8s.io/api/core/v1"
-	"k8s.io/client-go/dynamic"
-	kclient "k8s.io/client-go/kubernetes"
 )
 
 type TurboActionExecutorInput struct {
@@ -26,20 +25,18 @@ type TurboActionExecutor interface {
 }
 
 type TurboK8sActionExecutor struct {
-	kubeClient    *kclient.Clientset
-	dynamicClient dynamic.Interface
-	cApiClient    *clientset.Clientset
-	podManager    util.IPodManager
-	ormClient     *resourcemapping.ORMClient
+	clusterScraper *cluster.ClusterScraper
+	cApiClient     *clientset.Clientset
+	podManager     util.IPodManager
+	ormClient      *resourcemapping.ORMClient
 }
 
-func NewTurboK8sActionExecutor(kubeClient *kclient.Clientset, dynamicClient dynamic.Interface, cApiClient *clientset.Clientset,
+func NewTurboK8sActionExecutor(clusterScraper *cluster.ClusterScraper, cApiClient *clientset.Clientset,
 	podManager util.IPodManager, ormSpec *resourcemapping.ORMClient) TurboK8sActionExecutor {
 	return TurboK8sActionExecutor{
-		kubeClient:    kubeClient,
-		dynamicClient: dynamicClient,
-		cApiClient:    cApiClient,
-		podManager:    podManager,
-		ormClient:     ormSpec,
+		clusterScraper: clusterScraper,
+		cApiClient:     cApiClient,
+		podManager:     podManager,
+		ormClient:      ormSpec,
 	}
 }