--- a/pkg/discovery/monitoring/master/config.go
+++ b/pkg/discovery/monitoring/master/config.go
@@ -3,19 +3,15 @@ package master
 import (
 	"github.com/turbonomic/kubeturbo/pkg/cluster"
 	"github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/types"
-	"k8s.io/client-go/dynamic"
-
-	"k8s.io/client-go/kubernetes"
 )
 
 type ClusterMonitorConfig struct {
 	clusterInfoScraper *cluster.ClusterScraper
 }
 
-func NewClusterMonitorConfig(kclient *kubernetes.Clientset, dynamicClient dynamic.Interface) *ClusterMonitorConfig {
-	k8sClusterScraper := cluster.NewClusterScraper(kclient, dynamicClient)
+func NewClusterMonitorConfig(clusterScraper *cluster.ClusterScraper) *ClusterMonitorConfig {
 	return &ClusterMonitorConfig{
-		clusterInfoScraper: k8sClusterScraper,
+		clusterInfoScraper: clusterScraper,
 	}
 }