JIRA: https://vmturbo.atlassian.net/browse/OM-61188

**Intent**:
In cluster monitoring, we invoke api call in every discover for pod owner info, which is actually very expensive. We should introduce cache for those metadata kind of information, to avoid repeated api call without new information.

**Implementation**:
1. In `cluster_info_scraper.go`, create a `cache`(`turbostore.ITurboCache` type) in `ClusterScraper` struct with expiration duration as `defaultCacheTTL` (24 hours). `TurboCache` is an existing struct which makes use of cache library https://github.com/patrickmn/go-cache. The cached data will be cleaned up after `defaultCacheTTL` (24 hours). Meanwhile the lock inside go-cache library is a `sync.RWMutex`, which is very efficient in the case where there are more reads than writes.
2. Move the main logic from `GetPodGrandInfo` func in `pod_util` to `GetPodGrandparentInfo` in `cluster_info_scraper`. And use cache data inside `GetPodGrandparentInfo` func if exists.
3. There are 2 places which call `GetPodGrandparentInfo` func: 1) `ClusterMonitor::getPodOwner` in `cluster_monitor` and 2)`newK8sControllerUpdaterViaPod` func in `k8s_controller_updater`. Refactor the code to use the same `clusterScraper` (with the same cache) in both places.

**Testing Done**:

1. After the changes, controller kind is still shown correctly in UI:
<img width="1148" alt="Screen Shot 2020-08-17 at 01 43 04" src="https://user-images.githubusercontent.com/23689754/90361006-1c39ff80-e02b-11ea-8152-84b915d3d799.png">
And action on WorkloadController can still be executed successfully.

---
2. As for performance improvement, used [pprof](https://golang.org/pkg/net/http/pprof/) tool to profile CPU performance by running kubeturbo of cluster `K8s 1.18.5 in DC11` locally

**Environment**:
macOS
CPU: 2.5 GHz Intel Core i7

Before and after changes run such command:
```
go tool pprof http://127.0.0.1:10265/debug/pprof/profile?seconds=10
```
and then trigger a rediscovery to profile the runtime of each functions during discovery.

**Evidence**:
By repeating the above steps multiple times,
- before the change, `ClusterMonitor::Do` func runs around 30ms, where getPodOwner takes almost the whole 30ms:
```
      flat  flat%   sum%        cum   cum%
         0     0%   100%       30ms  7.32%  github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/master.(*ClusterMonitor).Do
         0     0%   100%       30ms  7.32%  github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/master.(*ClusterMonitor).RetrieveClusterStat
         0     0%   100%       30ms  7.32%  github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/master.(*ClusterMonitor).findNodeStates
         0     0%   100%       30ms  7.32%  github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/master.(*ClusterMonitor).genNodePodsMetrics
         0     0%   100%       30ms  7.32%  github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/master.(*ClusterMonitor).genNodeResourceMetrics
         0     0%   100%       30ms  7.32%  github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/master.(*ClusterMonitor).getPodOwner
```

- after the change, `ClusterMonitor::Do` func runs much faster with no more than 10 ms:
```
      flat  flat%   sum%        cum   cum%
         0     0%   100%       10ms  3.12%  github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/master.(*ClusterMonitor).Do
         0     0%   100%       10ms  3.12%  github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/master.(*ClusterMonitor).RetrieveClusterStat
         0     0%   100%       10ms  3.12%  github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/master.(*ClusterMonitor).findNodeStates
         0     0%   100%       10ms  3.12%  github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/master.(*ClusterMonitor).genNodePodsMetrics
         0     0%   100%       10ms  3.12%  github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/master.(*ClusterMonitor).genNodeResourceMetrics
         0     0%   100%       10ms  3.12%  github.com/turbonomic/kubeturbo/pkg/discovery/monitoring/master.(*ClusterMonitor).genOwnerMetrics
```
where `getPodOwner` func runs much faster than 10ms (looks the functions less than 10 ms are not shown in the profiling results)